[FLINK-6786] [metrics] Deduplicate QueryScopeIntoTest
This closes #4034.
