+++ flink-runtime/src/test/java/org/apache/flink/runtime/metrics/dump/QueryScopeInfoTest.java
@@ -20,11 +20,6 @@ package org.apache.flink.runtime.metrics.dump;
 
 import org.junit.Test;
 
-import static org.apache.flink.runtime.metrics.dump.QueryScopeInfo.INFO_CATEGORY_JM;
-import static org.apache.flink.runtime.metrics.dump.QueryScopeInfo.INFO_CATEGORY_JOB;
-import static org.apache.flink.runtime.metrics.dump.QueryScopeInfo.INFO_CATEGORY_OPERATOR;
-import static org.apache.flink.runtime.metrics.dump.QueryScopeInfo.INFO_CATEGORY_TASK;
-import static org.apache.flink.runtime.metrics.dump.QueryScopeInfo.INFO_CATEGORY_TM;
 import static org.junit.Assert.assertEquals;
 
 /**
@@ -32,46 +27,133 @@ import static org.junit.Assert.assertEquals;
  */
 public class QueryScopeInfoTest {
 	@Test
-	public void testJobManagerMetricInfo() {
-		QueryScopeInfo.JobManagerQueryScopeInfo info = new QueryScopeInfo.JobManagerQueryScopeInfo("abc");
-		assertEquals("abc", info.scope);
-		assertEquals(INFO_CATEGORY_JM, info.getCategory());
+	public void testJobManagerQueryScopeInfo() {
+		QueryScopeInfo.JobManagerQueryScopeInfo info = new QueryScopeInfo.JobManagerQueryScopeInfo();
+		assertEquals(QueryScopeInfo.INFO_CATEGORY_JM, info.getCategory());
+		assertEquals("", info.scope);
+
+		info = info.copy("world");
+		assertEquals(QueryScopeInfo.INFO_CATEGORY_JM, info.getCategory());
+		assertEquals("world", info.scope);
+
+		info = new QueryScopeInfo.JobManagerQueryScopeInfo("hello");
+		assertEquals(QueryScopeInfo.INFO_CATEGORY_JM, info.getCategory());
+		assertEquals("hello", info.scope);
+
+		info = info.copy("world");
+		assertEquals(QueryScopeInfo.INFO_CATEGORY_JM, info.getCategory());
+		assertEquals("hello.world", info.scope);
 	}
 
 	@Test
-	public void testTaskManagerMetricInfo() {
-		QueryScopeInfo.TaskManagerQueryScopeInfo info = new QueryScopeInfo.TaskManagerQueryScopeInfo("tmid", "abc");
-		assertEquals("abc", info.scope);
+	public void testTaskManagerQueryScopeInfo() {
+		QueryScopeInfo.TaskManagerQueryScopeInfo info = new QueryScopeInfo.TaskManagerQueryScopeInfo("tmid");
+		assertEquals(QueryScopeInfo.INFO_CATEGORY_TM, info.getCategory());
+		assertEquals("", info.scope);
 		assertEquals("tmid", info.taskManagerID);
-		assertEquals(INFO_CATEGORY_TM, info.getCategory());
+
+		info = info.copy("world");
+		assertEquals(QueryScopeInfo.INFO_CATEGORY_TM, info.getCategory());
+		assertEquals("world", info.scope);
+		assertEquals("tmid", info.taskManagerID);
+
+		info = new QueryScopeInfo.TaskManagerQueryScopeInfo("tmid", "hello");
+		assertEquals(QueryScopeInfo.INFO_CATEGORY_TM, info.getCategory());
+		assertEquals("hello", info.scope);
+		assertEquals("tmid", info.taskManagerID);
+
+		info = info.copy("world");
+		assertEquals(QueryScopeInfo.INFO_CATEGORY_TM, info.getCategory());
+		assertEquals("hello.world", info.scope);
+		assertEquals("tmid", info.taskManagerID);
 	}
 
 	@Test
-	public void testJobMetricInfo() {
-		QueryScopeInfo.JobQueryScopeInfo info = new QueryScopeInfo.JobQueryScopeInfo("jobid", "abc");
-		assertEquals("abc", info.scope);
+	public void testJobQueryScopeInfo() {
+		QueryScopeInfo.JobQueryScopeInfo info = new QueryScopeInfo.JobQueryScopeInfo("jobid");
+		assertEquals(QueryScopeInfo.INFO_CATEGORY_JOB, info.getCategory());
+		assertEquals("", info.scope);
 		assertEquals("jobid", info.jobID);
-		assertEquals(INFO_CATEGORY_JOB, info.getCategory());
+
+		info = info.copy("world");
+		assertEquals(QueryScopeInfo.INFO_CATEGORY_JOB, info.getCategory());
+		assertEquals("world", info.scope);
+		assertEquals("jobid", info.jobID);
+
+		info = new QueryScopeInfo.JobQueryScopeInfo("jobid", "hello");
+		assertEquals(QueryScopeInfo.INFO_CATEGORY_JOB, info.getCategory());
+		assertEquals("hello", info.scope);
+		assertEquals("jobid", info.jobID);
+
+		info = info.copy("world");
+		assertEquals(QueryScopeInfo.INFO_CATEGORY_JOB, info.getCategory());
+		assertEquals("hello.world", info.scope);
+		assertEquals("jobid", info.jobID);
 	}
 
 	@Test
-	public void testTaskMetricInfo() {
-		QueryScopeInfo.TaskQueryScopeInfo info = new QueryScopeInfo.TaskQueryScopeInfo("jid", "vid", 2, "abc");
-		assertEquals("abc", info.scope);
-		assertEquals("jid", info.jobID);
-		assertEquals("vid", info.vertexID);
+	public void testTaskQueryScopeInfo() {
+		QueryScopeInfo.TaskQueryScopeInfo info = new QueryScopeInfo.TaskQueryScopeInfo("jobid", "taskid", 2);
+		assertEquals(QueryScopeInfo.INFO_CATEGORY_TASK, info.getCategory());
+		assertEquals("", info.scope);
+		assertEquals("jobid", info.jobID);
+		assertEquals("taskid", info.vertexID);
 		assertEquals(2, info.subtaskIndex);
-		assertEquals(INFO_CATEGORY_TASK, info.getCategory());
+
+		info = info.copy("world");
+		assertEquals(QueryScopeInfo.INFO_CATEGORY_TASK, info.getCategory());
+		assertEquals("world", info.scope);
+		assertEquals("jobid", info.jobID);
+		assertEquals("taskid", info.vertexID);
+		assertEquals(2, info.subtaskIndex);
+
+		info = new QueryScopeInfo.TaskQueryScopeInfo("jobid", "taskid", 2, "hello");
+		assertEquals(QueryScopeInfo.INFO_CATEGORY_TASK, info.getCategory());
+		assertEquals("hello", info.scope);
+		assertEquals("jobid", info.jobID);
+		assertEquals("taskid", info.vertexID);
+		assertEquals(2, info.subtaskIndex);
+
+		info = info.copy("world");
+		assertEquals(QueryScopeInfo.INFO_CATEGORY_TASK, info.getCategory());
+		assertEquals("hello.world", info.scope);
+		assertEquals("jobid", info.jobID);
+		assertEquals("taskid", info.vertexID);
+		assertEquals(2, info.subtaskIndex);
 	}
 
 	@Test
-	public void testOperatorMetricInfo() {
-		QueryScopeInfo.OperatorQueryScopeInfo info = new QueryScopeInfo.OperatorQueryScopeInfo("jid", "vid", 2, "opname", "abc");
-		assertEquals("abc", info.scope);
-		assertEquals("jid", info.jobID);
-		assertEquals("vid", info.vertexID);
+	public void testOperatorQueryScopeInfo() {
+		QueryScopeInfo.OperatorQueryScopeInfo info = new QueryScopeInfo.OperatorQueryScopeInfo("jobid", "taskid", 2, "opname");
+		assertEquals(QueryScopeInfo.INFO_CATEGORY_OPERATOR, info.getCategory());
+		assertEquals("", info.scope);
+		assertEquals("jobid", info.jobID);
+		assertEquals("taskid", info.vertexID);
 		assertEquals("opname", info.operatorName);
 		assertEquals(2, info.subtaskIndex);
-		assertEquals(INFO_CATEGORY_OPERATOR, info.getCategory());
+
+		info = info.copy("world");
+		assertEquals(QueryScopeInfo.INFO_CATEGORY_OPERATOR, info.getCategory());
+		assertEquals("world", info.scope);
+		assertEquals("jobid", info.jobID);
+		assertEquals("taskid", info.vertexID);
+		assertEquals("opname", info.operatorName);
+		assertEquals(2, info.subtaskIndex);
+
+		info = new QueryScopeInfo.OperatorQueryScopeInfo("jobid", "taskid", 2, "opname", "hello");
+		assertEquals(QueryScopeInfo.INFO_CATEGORY_OPERATOR, info.getCategory());
+		assertEquals("hello", info.scope);
+		assertEquals("jobid", info.jobID);
+		assertEquals("taskid", info.vertexID);
+		assertEquals("opname", info.operatorName);
+		assertEquals(2, info.subtaskIndex);
+
+		info = info.copy("world");
+		assertEquals(QueryScopeInfo.INFO_CATEGORY_OPERATOR, info.getCategory());
+		assertEquals("hello.world", info.scope);
+		assertEquals("jobid", info.jobID);
+		assertEquals("taskid", info.vertexID);
+		assertEquals("opname", info.operatorName);
+		assertEquals(2, info.subtaskIndex);
 	}
 }
